chore: refactor integration tests to extract any_charm into a file
